requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
defusedxml==0.7.0rc1
Django==3.0.6
django-bootstrap4==1.1.1
django-cleanup==4.0.0
django-cors-headers==3.4.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
djangorestframework==3.11.1
easy-thumbnails==2.7
idna==2.10
oauthlib==3.1.0
Pillow==7.1.2
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
